App code reorganization
