+++ app/src/main/java/com/topjohnwu/magisk/arch/BaseViewModel.kt
@@ -13,12 +13,12 @@ import androidx.navigation.NavDirections
 import com.topjohnwu.magisk.BR
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.core.Info
+import com.topjohnwu.magisk.databinding.ObservableHost
+import com.topjohnwu.magisk.databinding.set
 import com.topjohnwu.magisk.events.BackPressEvent
 import com.topjohnwu.magisk.events.NavigationEvent
 import com.topjohnwu.magisk.events.PermissionEvent
 import com.topjohnwu.magisk.events.SnackbarEvent
-import com.topjohnwu.magisk.utils.ObservableHost
-import com.topjohnwu.magisk.utils.set
 import kotlinx.coroutines.Job
 
 abstract class BaseViewModel(
