+++ app/src/main/java/com/topjohnwu/magisk/core/utils/Keygen.kt
@@ -5,9 +5,9 @@ import android.content.pm.PackageManager
 import android.util.Base64
 import android.util.Base64OutputStream
 import com.topjohnwu.magisk.core.Config
-import com.topjohnwu.signing.CryptoUtils.readCertificate
-import com.topjohnwu.signing.CryptoUtils.readPrivateKey
-import com.topjohnwu.signing.KeyData
+import com.topjohnwu.magisk.signing.CryptoUtils.readCertificate
+import com.topjohnwu.magisk.signing.CryptoUtils.readPrivateKey
+import com.topjohnwu.magisk.signing.KeyData
 import org.bouncycastle.asn1.x500.X500Name
 import org.bouncycastle.cert.jcajce.JcaX509CertificateConverter
 import org.bouncycastle.cert.jcajce.JcaX509v3CertificateBuilder
