+++ app/src/main/java/com/topjohnwu/magisk/signing/ApkSignerV2.java
@@ -1,4 +1,4 @@
-package com.topjohnwu.signing;
+package com.topjohnwu.magisk.signing;
 
 import java.nio.BufferUnderflowException;
 import java.nio.ByteBuffer;
