+++ app/src/main/java/com/topjohnwu/magisk/signing/JarMap.java
@@ -1,4 +1,4 @@
-package com.topjohnwu.signing;
+package com.topjohnwu.magisk.signing;
 
 import java.io.Closeable;
 import java.io.File;
