+++ app/src/main/java/com/topjohnwu/magisk/ui/hide/HideRvItems.kt
@@ -7,9 +7,9 @@ import com.topjohnwu.magisk.BR
 import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.databinding.ComparableRv
 import com.topjohnwu.magisk.databinding.ObservableDiffRvItem
+import com.topjohnwu.magisk.databinding.addOnPropertyChangedCallback
+import com.topjohnwu.magisk.databinding.set
 import com.topjohnwu.magisk.ktx.startAnimations
-import com.topjohnwu.magisk.utils.addOnPropertyChangedCallback
-import com.topjohnwu.magisk.utils.set
 import com.topjohnwu.superuser.Shell
 import kotlin.math.roundToInt
