+++ app/src/main/java/com/topjohnwu/magisk/ui/inflater/LayoutInflaterFactory.kt
@@ -1,7 +1,6 @@
 package com.topjohnwu.magisk.ui.inflater
 
 import android.content.Context
-import android.os.Build
 import android.util.AttributeSet
 import android.view.InflateException
 import android.view.LayoutInflater
