+++ app/src/main/java/com/topjohnwu/magisk/ui/superuser/PolicyRvItem.kt
@@ -7,7 +7,7 @@ import com.topjohnwu.magisk.R
 import com.topjohnwu.magisk.core.model.su.SuPolicy
 import com.topjohnwu.magisk.databinding.ObservableDiffRvItem
 import com.topjohnwu.magisk.databinding.RvContainer
-import com.topjohnwu.magisk.utils.set
+import com.topjohnwu.magisk.databinding.set
 
 class PolicyRvItem(
     override val item: SuPolicy,
